backend/requirements.txt
Django>=3.2.5
djangorestframework>=3.12.2
psycopg2-binary>=2.8
dj-database-url>=0.5.0
asgiref>=3.3.1
pytz>=2020.5
sqlparse>=0.4.1
django-cors-headers==3.7.0
djangorestframework-simplejwt>=4.6.0